First changes to make the cardproxy less smart.
The refactoring on the irma webservice has sparkled the realization that a too smart cardproxy app makes for a lot of duplicate work at the webservice and cardproxy app. The changes in this commit work towards an approach where the javascript in the browser is the "master" and the cardproxy app behaves more like a "dumb" cardreader. This has the additional advantage that it makes the user experience more consistent across smartcard reader devices.

Do note however that the current version is incomplete and mostly untested. I.e., it is work in progress.